src/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
psycopg2-binary
